Fix error on reading EOF
Valgrind gave us errors that a "Conditional jump or move depends on
uninitialised value(s)". This happened because we received an EOF when
reading in the next token, however we immediately call a switch on
this uninitialised input character. Adding a check and returning a
default token on input error solves this issue.